Use the accurate line/column numbers to get accurate ranges.
CoffeeScriptRedux has a preprocessing step that inserts control
characters to ensure a PEG-generated parser can properly parse the
source. Unfortunately this means that the character offset during
parsing includes these control characters, and the offset is used to
create the range property on the AST nodes.

The line and column numbers, however, are accurate because of a hack in
the parser that forces the control characters to be ignored. The
workaround this commit adds is to create a mapping of line/column
values to offsets and use that to replace the bad offsets provided by
CSR.

See https://github.com/michaelficarra/CoffeeScriptRedux/issues/335.