Be more aggressive about fixing ranges.
Just because the raw data matches the string at the claimed range
doesn’t mean that’s the right range. For example, in the test added it
just so happens that the `e` in the `else if` condition has an
incorrect range that points it to the “e” in the`else` following the
block for the `else if`. This was causing parentheses to be inserted
incorrectly.