Add a preprocessor to handle POST-if and POST-unless.
Preprocessing steps will allow us to make modifications to the original
source that convert the original CoffeeScript into a form that is more
amenable to transformation to JavaScript.

In this case, we cannot use MagicString to turn `return if a is b` into
`if (a === b) { return; }` because it requires swapping the condition
and consequent *and* re-writing the condition. Doing so violates
MagicString’s precondition that you can only modify a character at a
given index once. Instead, we turn `return if a is b` into `if a is b
then return` and re-parse it so that each character position is only
re-written once in the transformation to JavaScript.