Add separate variable declarations for assignments-as-expressions.
When an assignment is used as a block statement we can simply turn the
assignment into a variable declaration, e.g.

  a = 1 // may become `var a = 1;`

However, assignments are allowed anywhere expressions are and therefore
may not always be suitable for conversion to variable declarations, as
they are statements. In such cases we must pre-declare the variables
being assigned, e.g. `a(b = 1)` must create a `var b;` with no
initializer:

  var b;
  a(b = 1);