Ensure back-to-back escapes work correctly.
This corrects a logic error where we were skipping 2 characters when encountering an escape character. This incorrectly assumed that those two characters would be the escape character (i.e. `\`) and the character following it, but in reality the call to `consume` already skipped the escape character, so skipping two more skipped the escaped character and the character after it.

Fixes #104.