Improve expression vs. statement generation.
Since so many things in CoffeeScript are expressions that cannot be in JavaScript, we have to be able to decide which style to generate. For example, `throw` and `if` can be used as expressions in CoffeeScript.  These must either be wrapped in IIFEs or, for `if`, turned into ternary or `do` expressions. Sometimes both ways would be valid (i.e. do we turn `=> if a then b` into `() => a ? b : undefined` or `() => { if (a) { return b; } }`? We need a system like this so parent nodes can express their preference to their children and the children can override if they decide it’s better to do something different.