Fine-tune insertion positions.
This renames `start` and `end` to `contentStart` and `contentEnd`, and `before` and `after` to `outerStart` and `outerEnd` to denote that they are outside the surrounding parentheses. This also adds two new ones, `innerStart` and `innerEnd` which are for inserting inside the surrounding parentheses.

Perhaps the biggest functional change is that these are not longer fixed as soon as the patchers are constructed, but can change based on edits. If a patcher edits outside its bounds then the bounds will change for subsequent edits. This ensures that trailing punctuation, such as `;` or `)`, ends up at the right place.