refactor: improve parentheses handling for binary operators
This causes decaffeinate to insert parentheses in places where it wouldn't have before. In particular, nested binary operators will now be wrapped in parentheses. This is a side effect of accepting a `needsParens` option in `BinaryOperator`'s `patchAsExpression` method which we'll need when moving binary operators into a location that they will need parentheses to be unambiguous.