feat: strip trailing `return` statements without an argument
The NormalizeStage sometimes adds them to prevent inserted statements from being implicitly returned unintentionally. This has the side effect of stripping empty trailing return statements that were in the original source code, but those were probably also there to prevent an implicit return and so would not normally be in the JavaScript version of the code anyway.

Closes #193.