fix: don't remove comments when removing trailing empty returns
Before this commit, the comment after a removed return statement would be
destroyed:
http://decaffeinate-project.org/repl/#?evaluate=true&code=-%3E%0A%20%20a%20%20%23%20foo%0A%20%20return

Now, the code is a little more careful by deleting up to the start of the line
or the previous semicolon. Also, the code now exits early and doesn't patch the
`return` node or insert a semicolon at the end, since that was causing a
semicolon at the end of the comment.

This also fixes a crash on this type of input that I ran into when using
magic-string with  https://github.com/Rich-Harris/magic-string/pull/89 applied.