feat: handle expansions in array assignment
This is a partial implementation of expansions, which handles them for array
assignment and destructuring, but not for function parameters or nested within
more complicated assignments. For those unsupported cases, I throw an error
linking to #268 when the expansion node is traversed.

Currently the only special case is `[a, b, ...]`. Every other case just turns
into a series of assignments that do array lookups.