feat: handle some "for in" loop expressions using filter and map (#321)
Partial fix for #156.

This commit handles "for in" using the strategy of transforming it to
`a.filter(...).map(...)`. This works in the most common cases, but there are some
edge cases that it doesn't handle (where I throw an error):
* Loop expressions using a `by` clause.
* Loop expressions where the body cannot be patched as an expression. In
  particular, I don't think there's a great way to handle `break`, `continue`,
  and `return` within loop expressions using `map` and a lambda.
* Expressions that use a filter and also assign the index to a variable. In this
  case, the code would be incorrect, because the filter step changes the size of
  the array and thus the indexes.

This technically could change behavior when looping over array-like objects,
like `{length: 2, "0": "hello", "1", "world"}`. But I think that's obscure
enough that it's unlikely to come up in practice, and if it does, it's better to
generate the nice code by default and the 100% correct code as an option.

I also added some skipped tests for cases that this commit doesn't support.

Out of the 180 files in the codebase I'm working with that were failing due to
for loop expressions, this commit fixes 159 of them. Of the remaining 21 of
them, 11 of them fail due to a non-expression body (about half of which had
`continue` statements), 4 of them fail due to a "for of" expression, 4 of them
fail because they use a "by" clause, and 2 of them fail because they use
"for own".