fix: improve the pacement of the close-paren for implicit function calls
This fixes a crash where a close-paren was being placed in the wrong place.
[(repl)](http://decaffeinate-project.org/repl/#?evaluate=true&code=%5Ba(b%2C%20c%20%7B%0A%7D)%5D)

The previous code was using `appendLineAfter`, which was using
`getBoundingPatcher` as a way of determining where it's safe to place the
paren. However, `getBoundingPatcher` doesn't seem to always do that, for example
when the node is an argument to a function taking multiple arguments. Instead, I
now just always place the close-paren in the function call's `innerEnd`. Also,
when deciding whether to put the close-paren on a new line or right after the
end of the call, I now put it right at the end if the call ends with a `}` or a
`]`, which fixes a bunch of ugly newlines that I was seeing.

Unfortunately, this causes a comment to be placed one line below the ideal spot
in one place, but that seems acceptable for now, especially since the comment is
still preserved.

Maybe a longer-term approach would be to improve `getBoundingPatcher` and
`appendLineAfter`, but at least, for now, I think this change makes the code
look better and makes it less likely for mismatched parens to pop up.