chore: update CoffeeScript dependency to the decaffeinate fork
This doesn't actually fix anything because this dependency is only used for a
test util that doesn't need the patch, but it seems like a good idea for
consistency, and in case we want to use it elsewhere.

The actual fix will be pulled in with the update to decaffeinate-parser.