fix: improve formatting in IIFE-style while loop expressions (#390)
Progress toward #156.

The existing while loop IIFE code was trying to cram all of the generated code
on one line, which looked kind of bad in longer cases. Since I'm going to be
extending this code to all types of loops soon, I figured it would be good to
improve the formatting first.

The new rule is that if a loop is formatted within an IIFE, we never treat it
as an inline node. To help support this type of case, I added a `setIndent`
method that forces the indentation level to a certain value, which helps ensure
that the generated code is nicely nested regardless of the details of how the
original code was formatted.

Unfortunately, the new indentation caused some excessive whitespace in some
`yield` examples, and there are some cases where newlines in the original
CoffeeScript cause blank lines in the generated JS. But I think those issues are
all relatively minor, and can be fixed later.