feat: add support for all for loops used as expressions (#392)
This changes `for` loops used as expressions to use a statement-style loop
within an IIFE, unless we already have a better way to transform the expression.

Conceptually, this was fairly easy since we already do this type of thing for
`while` loops, so it was mostly a matter of pulling that logic into a shared
`LoopPatcher` class. However, getting the formatting right in all cases ended up
being a huge pain, so I ended up reworking how a lot of that stuff works.

There are two main things that I introduced:
* All styles of loops now use common terminology for three indentation levels:
  the top-level loop where the `for` or `while` goes, the first level of the
  loop body, where we often generate an assignment statement, and the "main"
  loop body, which might be indented one more level to be wrapped in a generated
  `if` statement. Various code paths can now reference those logical indentation
  levels.
* To handle the three different cases of prepending a line to the front of a
  block, I added a new `insertLineBefore` method. In particular, this handles
  both the awkward inline-to-non-inline case, and also allows prepending lines
  that are at a lower indentation level than the block itself.

One important property of the implementation here is that the only indentation
is in the loop body, and the indentation happens exactly once. I tried lots of
approaches involving indenting twice or indenting two different nodes, and all
of those seemed to confuse magic-string in various edge cases. Fewer operations
is better in this case, it seems.

Closes #156.