feat: allow code within class bodies
Closes #397. See that issue for a much more thorough description of the strategy
here. This change doesn't handle all cases described in that issue (in
particular, it doesn't support arbitrary code within classes that are used in an
expression context), but it handles enough cases to support all classes in the
codebase I'm working with.

Since this transform significantly rearranges classes, I decided to do it as a
CS->CS transform in the normalize stage. That should hopefully avoid weird
magic-string issues in the main stage.

This approach is maybe a bit controversial in that it changes the implementation
for class fields to use prototype properties. In my case, putting class
properties on the prototype is necessary to avoid widespread correctness issues
in my codebase. (See http://benmccormick.org/2015/07/06/backbone-and-es6-classes-revisited/
for more details.) My current thought is I'll leave the class fields code in
place and we can make the behavior configurable with a command line flag if
there's demand for it. Or I can add the flag as part of this PR.