fix: allow a space in logical OR assignments (#424)
Closes #423.

CoffeeScript allows things like `or =` and `|| =` and treats them as compound
assignments (even though in general it doesn't allow spaces in operators; for
example, `? =` is NOT a valid assignment operator). Our check wasn't handling
that case, so `isOrOp` was returning `false` and the assignment was being
treated as a logical AND. To fix, I just looked at the start of the operator
rather than the whole operator.