fix: allow parameter property assignments with defaults (#426)
Fixes #232.

The code for handling `@` in params already had a check to make sure it doesn't
traverse up the right side of a default param assignment (otherwise `(a = @b) ->`
would get converted incorrectly), but it also skipped the valid case where we
traversed up the left side. To fix, I just needed to see which of the parent's
children the current patcher is when traversing up.