fix: correctly handle destructure parameters with property assignments
Closes #213.

In a case like `{@a}`, the decaffeinate-parser AST uses the same node for the
key and the value, so while searching for this-assignment nodes we were seeing
`@a` twice (even though conceptually it might be better represented as `{a: @a}`
in the AST, but that causes problems elsewhere). This meant we were adding two
assignment statements when we should have been adding one. There was already a
mechanism to ignore some nodes in parameter bodies (used to ignore usages of `@`
within parameter defaults), so this issue can be fixed by just ignoring object
keys, which always just say the name of the field to be extracted rather than
having any effect on the assignments that are done.