fix: give a more clear error message on complex conditionals
Previously, we would just skip patching IIFE-style conditions, causing a
confusing error later, but now we give a helpful message pointing at a bug.

I also made it so we require both the consequent and the alternate (if present)
to be patchable as an expression before we say that the condition can be a
ternary; otherwise, it was hitting a more opaque error in some cases.