feat: implement IIFE-style conditionals (#444)
Closes #388.

When an `if` expression is complicated enough, we can't reasonable turn it into
a ternary; it's cleaner (and sometimes necessary) to use an IIFE instead.
We already have IIFE code working for loops, so this borrows most of the
implementation details from that.

I didn't focus as much on formatting as for the loop case, so it's certainly
possible that there are some situations where the indentation looks wrong, but
in the few examples I tried it looks pretty good. If there are formatting
problems, those can be fixed as lower-priority follow-up issues.

I also made it so we require both the consequent and the alternate (if present)
to be patchable as an expression before we say that the condition can be a
ternary; otherwise we would sometimes attempt a ternary patch that would then
fail.