fix: when patching an arg, don't write past the end of the call (#445)
Closes #412
Closes #410
Closes #405
Closes #379
Closes #370
Closes #276

The `appendLineAfter` method was careful to never insert code past the end of
the current node's "bounding patcher", but the bounding patcher was computed in
a way that didn't work in all cases; it just found the lowest enclosing node
that's surrounded by parens. So in `f(a, b)`, the bounding patcher for `b` is
the whole expression, so there's nothing to stop `b` from inserting code past
the end of the close-paren. To fix, I made the bounding patcher code a little
smarter to recognize when it's traversing up a function call node and stop
there.