fix: properly convert for expressions with implicit object bodies
Closes #467.

This was a magic string issue where inserts were happening out of order. Now,
when patching expression-style for-in loops, we patch the body in the middle,
and split the `surroundInParens` call into an insert before and after. Placing
the open-paren in a consistent way ended up being tricky, and required some
cases to handle both the inline and multiline cases.