fix: respect bounds when writing the end of a switch expression
Closes #448.

The switch patcher was using `appendToEndOfLine`, which was just writing to the
end of the current line, even if it was past the end of the expression (in this
case, it was writing past the close-paren for the function call). To fix, I just
made `appendToEndOfLine` include the same logic as `appendLineAfter`: write to
the right place if possible, but not past the bounding patcher.