fix: properly handle break and continue in loop expressions
Fixes #473

`break` and `continue` were parsed as regular identifiers, so the loop
expression patching code would sometimes make a lambda trying to return them or
generated code like `result.push(break)`. Probably the cleanest fix would be to
parse `break` and `continue` differently in the AST, but the easy fix is to just
say that the identifier can't be patched as an expression in those cases, and
handle that properly.