fix: improve formatting when a function is the last argument to a function (#465)
Closes #451.

There was already a special case that when a function expression is surrounded
by parens, the `}` is placed immediately before the close-paren. This also
worked for single-argument function invocations, but in a multi-argument
function invocation, neither argument is considered to be surrounded by parens,
so it wasn't doing anything smart. To fix, I added a new special case: if a
function expression is the last argument to a function, place the `}` just
before the CALL_END token.