fix: properly handle conditionals with a null consequent (#469)
Closes #305.

This was basically just a matter of going through ConditionalPatcher in both the
normalize stage and the main stage and making sure they account for the case
where `consequent` is null. There are some regions of code that only happen when
`consequent` is non-null, like post-`if` patching and code with a `then` token,
so I left those as-is.

The formatting is a little weird for a missing consequent when the conditional
is used in an expression context, but that's probably a rare case and can be
cleaned up later.