fix: correctly pass `this` in soaked function applications (#464)
Closes #462.

This fixes the JS output for code like `a().b?()` and `a()[b()]?()`, where the
`this` value needs to be set properly for the function call. The solution is
among the same lines as `__guard__`/`__guardFunc__`, but even uglier. Rather
than using a function to eager-evaluate the base and optionally call a lambda,
we eagerly evaluate both the object and the method name and then optionally pass
both into the lambda. There were also some details in handling cases like
`@a?()`, `a?.b?()`, and `a?.b.c?()`.

As usual, this (mostly) handles the crazy cases, and #336 is the task for making
the output nicer in the non-crazy cases. In the meantime, I always just manually
fix up the generated JS code to never use `__guard__`-type things.

One decision I made is that I treat `a.b?()` as `a?.b?()`, which means generated
code could potentially succeed when it would crash in CoffeeScript. It seems
like this likely wouldn't be a problem in practice, and makes it possible to do
chained soaked operations without things needing to get much more complex.

This change also doesn't handle soaked prototype accesses like `a::?b` and
`a::b?()`, but those weren't working before anyway and seem especially obscure.
It should be reasonable to implement them in the future if necessary, I think.