fix: change post-if patching to use insert
Closes #479.
Closes #459.
Closes #419.
Closes #378.
Closes #334.

The post-if patcher was changed to use `move` in magic-string, but that ended up
messing with the order that things were written in. Now that magic-string
properly handles `remove`, it should be safe to go back to using `insert`, with
the general strategy being to insert fragments in order in the source code
whenever possible.