fix: improve the process of patching code for use elsewhere
Fixes #498

Previously, you might call `patch` on a sub-patcher, then use `slice` to get the
generated code, but this didn't always work out well, since we always do
`insertLeft` (for magic-string sanity), so slice would sometimes miss an
open-paren inserted at the start.

This commit changes the loop filter case to use this new method, but probably
other code paths should be moved over to it as well in the future.