fix: don't include parens when patching loop filters (#488)
Fixes #483.

Since the condition in a loop is already going to be surrounded by parens,
there's no need to add an extra set of parens. Also, it was causing a bug where
one of the parens was kept but the other one wasn't.

This doesn't really solve the general case; ideally there would be a way to get
the expression including the parens.