fix: use expression patching for existence assign ops with soaks
Fixes #512

Existence assignment ops assume that if we're patching as a statement, we can
remain a statement (and thus change to an `if` statement), but the `__guard__`
approach to soak operations assumes that an assignment can have parens wrapped
around it. To avoid this conflict, we can have the existence assignment code
look ahead to see if there's a soak operation, and just do an expression-style
patch in that case. Since this should be a relatively rare case, the main goal
here is to just make sure the result is correct.