fix: improve handling of loop targets to get the right code (#518)
Fixes #511.

The new `patchAndGetCode` method is more robust than the previous approach,
especially in cases like soaks that add code to the start. Also, the code was
overwriting the target even if it didn't need to be extracted, which caused
problems in some cases, so only overwrite it if it's going to be replaced by
something different.