fix: patch the assignee when generating JS for-of loops
Fixes #520

It turns out that this-assignment works in a for-of loop in JS, so we can just
convert `for @a in b` to `for (this.a of b)`. The for-of conversion always
happens when there's a value assignee and no key assignee, so we can just always
patch the value assignee.