fix: place implicit call close parens more carefully to avoid later parse errors (#525)
Fixes #417
Fixes #387

In some cases, the code path to insert the implicit close-paren in a function
call would create a new line with a properly-indented close-paren, but that
would confuse the CoffeeScript parser if the following token was a close-paren
at a greater indentation. Now, in that case, we put the close-paren next to the
following close-paren so the indentation can't get messed up.