fix: remove unnecessary commas in the normalize stage (#526)
Fixes #447.

CoffeeScript is not particularly consistent about when to allow trailing commas
in multiline argument lists, array literals, and object literals. If the last
part of the argument or element is an implicit function call, then they are
allowed, but if the function call has parens, then the trailing comma is not
allowed. This means that the normalize stage can make code incorrect in some
cases, so we need to remove these unnecessary parens.