fix: properly insert brackets around object keys when necessary
Fixes #411

In most cases, object keys in CoffeeScript are just regular strings or
identifiers, so we can just patch them normally. But for template strings and
multiline herestrings, we need to put brackets around the string since those
types of object keys aren't allowed in JS. Also, for the specific case of a
template literal key with only a single expression, it's useful to just convert
that expression directly to the computed key form.