fix: allow throw statements containing a newline
Fixes #503

Apparently CoffeeScript lets you put a newline between throw and the error,
while JavaScript doesn't. To fix, just remove that spacing if necessary. Note
that this can sometimes end up removing comments, so ideally we would find a
place to move them, but this seems like an unlikely place to put a comment in a
case that already seems relatively rare, so this is probably fine for now.