fix: properly insert break statements in switches with conditionals
Fixes #449

The switch case code was assuming that because the consequent was in an implicit
return context, the code would return and no break was necessary. However,
that's not always true; in a conditional, there may be missing code paths,
meaning that no implicit return happens, so we need to also make sure that all
code paths exist.