fix: properly set patching bounds for new and soaked function call operations
Fixes #501

This is similar to #445, but that change tested exactly for the
`FunctionApplication` node (we can't reference patchers because `NodePatcher`
is independent of stage). But really, this happens in any case where we have an
argument list, so we need to broaden it to include new operations and soaked
function applications.