fix: always patch the target in for-of loops
Fixes #538

In PR #518, I changed the `getTargetReference` call in for-of loops to be under
a conditional, which meant that in some cases (when the target was repeatable),
the target wouldn't be patched at all. To fix, we can just always patch the
target, but only sometimes overwrite the actual target code with the reference.