fix: properly account for slice ranges ending in 0
Fixes #524

In an expression like `a[b..-1]`, it's incorrect to use 0 as the second slice
arg, since that refers to the start of the string, not the end. Instead, we want
to just omit the second arg. In the more general case where the right side is a
variable, we need to add `|| undefined` so that if the result is 0, we
effectively pass no arg to the function.