feat: implement bound generator functions
Fixes #516

Generally bound generator functions can be converted to regular `function`
expressions with `.bind(this)` at the end, but there are various edge cases like
when the function is in an object or class body.

Also fix some issues with fat arrow functions referencing `arguments` not always
having `this` bound.