feat: use `for(;;)` style loops for some `for [a..b]` style loops
This commit does not use `for(;;)` when the loop has a dynamic range unless it also has an explicit step. That means `for [a..b]` will become a `for…of` loop. I added two skipped tests for turning them into `for(;;)` style, but it would require a fairly verbose loop header. However, that may still be better than requiring the `__range__` helper.