fix: properly handle code with Windows newlines or mixed newline style
Fixes #550

We convert the newline style to \n at the start, then convert it to its original
style at the end. That way, decaffeinate can detect and generate \n newlines
without having to worry about the different styles.

There are some libraries out there for detecting and converting newline styles,
but they're a bit more complicated than I need and don't behave exactly the way
I want, and both operations are pretty easy to implement.

This change also means that we can simplify some of the code that was handling
CR characters. Generally, any code now can assume that '\n' is the only type of
newline.

Note that I don't make any attempt at supporting CR newlines. Hopefully Mac OS 9
is far enough in the past that we don't need to have to think about it.