fix: in switch cases, defer negation until just before patching (#563)
Fixes #562

The switch case patcher does a magic-string remove operation from the start of
the `when` token to the start of the condition. When the condition was negated,
it would sometimes have a `!` inserted at the start that would then be removed
by this remove operation, producing incorrect code. To fix, we can just track
the negation and delay it until just before the conditions are patched.

Generally, this moves code toward the best practice of making sure magic-string
operations happen in order, and a similar strategy may be useful if other
patchers cause issues.